Refactor LLM plugins. (#389)
