+++ ads/llm/__init__.py
@@ -5,14 +5,7 @@
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
 
 
-from ads.llm.langchain.plugins import GenerativeAI
-from ads.llm.langchain.plugins import OCIModelDeploymentTGI
+from ads.llm.langchain.plugins.llm_gen_ai import GenerativeAI
+from ads.llm.langchain.plugins.llm_md import ModelDeploymentTGI
 
 # from ads.llm.langchain.plugins import OCIModelDeploymentVLLM
-
-
-__all__ = [
-    "GenerativeAI",
-    "OCIModelDeploymentTGI",
-    # "OCIModelDeploymentVLLM",
-]
